requirements.txt
phe>=1.5.0
pillow>=9.0.0
numpy>=1.21.0
requests>=2.27.0
